chore: Reorganize project files
