+++ IssueTracker15/IssueTracker15.xcodeproj/xcshareddata/xcschemes/IssueTracker15.xcscheme
@@ -42,16 +42,6 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "1B87F719285C38FD002B47B7"
-               BuildableName = "ViewModelUnitTest.xctest"
-               BlueprintName = "ViewModelUnitTest"
-               ReferencedContainer = "container:IssueTracker15.xcodeproj">
-            </BuildableReference>
-         </TestableReference>
-         <TestableReference
-            skipped = "NO">
-            <BuildableReference
-               BuildableIdentifier = "primary"
                BlueprintIdentifier = "8D2113772859E5FD00B65676"
                BuildableName = "ContainerUnitTest.xctest"
                BlueprintName = "ContainerUnitTest"
