--- IssueTracker15/IssueTracker15/Test/UseCases/UseCaseModel.swift
@@ -1,8 +1,0 @@
-//
-//  UseCaseModel.swift
-//  IssueTracker15
-//
-//  Created by 백상휘 on 2022/06/17.
-//
-
-import Foundation
