--- IssueTracker15/IssueTracker15/Test/ViewTypes/BoldStringAdaptable.swift
@@ -1,9 +1,0 @@
-//
-// Created by 백상휘 on 2022/06/14.
-//
-
-import UIKit
-
-protocol BoldStringAdaptable: ViewTextBindable {
-    func bind(_ text: String, with font: UIFont)
-}
