--- IssueTracker15/IssueTracker15/Test/ViewTypes/LargeStringAdaptable.swift
@@ -1,9 +1,0 @@
-//
-// Created by 백상휘 on 2022/06/14.
-//
-
-import UIKit
-
-protocol LargeStringAdaptable: ViewTextBindable {
-    func bind(_ text: String, with font: UIFont)
-}
