+++ IssueTracker15/IssueTracker15/ViewBind/ViewBinding.swift
@@ -1,0 +1,12 @@
+//
+//  ViewBinding.swift
+//  IssueTracker15
+//
+//  Created by 백상휘 on 2022/06/27.
+//
+
+import Foundation
+
+protocol ViewBinding {
+    func inputViewEvent(_ target: ViewBindable, _ param: Any?)
+}
